[#6469] move infrastructure validation to dedicated class
